Made minor improvements in code quality.
Previously the code did not make use of some python features and did
joining and list comprehension manually. I've improved those areas that
needed it.

I've also simplified the window code to compute an upper and lower bounds,
and improved it by catching the entire window.

Example:
"This is a string."

Previously a window with a num_context of 2 would arrive on "is" and
generate a window of ["is", "a", "string."]. It now generates a window of
["This", "is", "a", "string"].

This should improve the overall accuracy, but especially at the beginning
and end of a card.

Finally I have removed all trailing whitespace.